Friends type is moved to shared typesFolder, removed mockTypes from chatContainer
