+++ client/src/types/User.ts
@@ -3,3 +3,14 @@ export interface User {
     username: string
     email: string
 }
+
+export interface PublicUser {
+    id: number
+    username: string
+    avatar: string | null
+}
+
+export interface Friendship {
+    friend: PublicUser
+    created_at: string
+}
